+++ gatling-charts-highcharts/src/main/scala/com/excilys/ebi/gatling/highcharts/series/ScatterSeries.scala
@@ -7,7 +7,7 @@ package com.excilys.ebi.gatling.highcharts.series
 
 import com.excilys.ebi.gatling.charts.series.Series
 
-class ScatterSeries(name: String, data: Seq[(Int, Long)], color: String) extends Series[Int, Long](name, data, List(color)) {
+class ScatterSeries(name: String, data: Seq[(Long, Long)], color: String) extends Series[Long, Long](name, data, List(color)) {
 
 	def elements: Seq[String] =
 		if (sample.isEmpty)
